twitch_mobile_framework/requirements.txt
pytest==7.4.3
selenium==4.15.2
webdriver-manager==4.0.2
pytest-html==4.1.1
pytest-xdist==3.3.1
allure-pytest==2.13.2
python-dotenv==1.0.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
Pillow==10.1.0
